+++ mopinion/dataclasses.py
@@ -1,8 +1,10 @@
-import re
-
-from typing import Optional, Union
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+from dataclasses import field
 from mopinion import settings
+from typing import Optional
+from typing import Union
+
+import re
 
 
 __all__ = [
@@ -44,8 +46,8 @@ class EndPoint(Argument):
             r"/datasets/\d+$",
             r"/datasets/\d+/fields$",
             r"/reports/\d+/fields$",
-            r"/datasets/\d+/feedback/\w+$",
-            r"/reports/\d+/feedback/\w+$",
+            r"/datasets/\d+/feedback$",
+            r"/reports/\d+/feedback$",
             r"/reports/\d+$",
             r"/reports$",
         ]
@@ -57,10 +59,9 @@ class EndPoint(Argument):
 @dataclass(frozen=True)
 class ApiRequestArguments(Argument):
     endpoint: EndPoint
-    version: str
     content_negotiation: str
     verbosity: str
-    method: str
+    version: str = None
 
     def __post_init__(self):
         # verbosity levels
@@ -70,19 +71,13 @@ class ApiRequestArguments(Argument):
                 f"'{', '.join(settings.VERBOSITY_lEVELS)}'"
             )
 
-        if self.method.lower() not in settings.ALLOWED_METHODS:
-            raise ValueError(
-                f"'{self.method}' is not a valid choice. Please consider one of: "
-                f"'{', '.join(settings.ALLOWED_METHODS)}'"
-            )
-
         if self.content_negotiation not in settings.CONTENT_NEGOTIATIONS:
             raise ValueError(
                 f"'{self.content_negotiation}' is not a valid content negotiation. "
                 f"Please consider one of: '{', '.join(settings.CONTENT_NEGOTIATIONS)}'"
             )
 
-        if self.version not in settings.VERSIONS:
+        if self.version and self.version not in settings.VERSIONS:
             raise ValueError(
                 f"'{self.version}' is not a valid version. Please consider one of: "
                 f"'{''.join(settings.VERSIONS)}'"
@@ -95,7 +90,6 @@ class ResourceUri(Argument):
     resource_name: str
     resource_id: Optional[Union[str, int]]
     sub_resource_name: Optional[str]
-    sub_resource_id: Optional[Union[str, int]]
 
     def __post_init__(self):
         from mopinion.client import MopinionClient
@@ -129,8 +123,6 @@ class ResourceUri(Argument):
             endpoint += f"/{self.resource_id}"
             if self.sub_resource_name:
                 endpoint += f"/{self.sub_resource_name}"
-                if self.sub_resource_id:
-                    endpoint += f"/{self.sub_resource_id}"
         self.endpoint = endpoint